--- src/_client-side.js
@@ -1,7 +1,0 @@
-import convert from './convert/index.js'
-import validate from './validate/index.js'
-import uncompress from './compress/uncompress.js'
-import reverse from './reverse/index.js'
-import classify from './classify/index.js'
-
-export { convert, uncompress, reverse, validate, classify }
