+++ src/again/03-reverseRules.js
@@ -1,0 +1,4 @@
+const revRules = function (pairs, model, threshold) {
+
+}
+export default revRules
